refactor : 모달창 수정
